added crease temp logo and login/sign up buttons
