+++ client/extension/calendar.png
