+++ client/extension/icon.png
