+++ client/extension/manifest.json
