+++ client/extension/materializeinitialize.js
