+++ client/extension/test-icon.png
