--- client/src/components/home.css
@@ -1,27 +1,0 @@
-
-body{
-    background-color: #FFFFFF;
-    margin: 0;
-
-}
-
-.homeTitle{
-
-}
-
-.hrBar{
-    border: solid #e5e5e5;
-    border-width: 2px 0 0;
-}
-
-.homeDivPart{
-    width: 100vw;
-    height: 70vmin;
-
-    border-bottom: 1px solid black;
-
-
-}
-.ggLogin{
-
-}
